requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Deprecated==1.2.9
docker==4.2.0
future==0.18.2
gitdb==4.0.2
github-repo==2019.2.23
GitPython==3.1.0
hupper==1.10.2
idna==2.9
ini2dict==2019.4.13
PasteDeploy==2.1.0
plaster==1.0
plaster-pastedeploy==0.7
psutil==5.7.0
public==2019.4.13
PyGithub==1.47
PyJWT==1.7.1
pyngrok==2.1.3
pyramid==1.10.4
PyYAML==5.3.1
requests==2.23.0
runcmd==2019.4.13
six==1.14.0
smmap==3.0.1
translationstring==1.3
urllib3==1.25.8
values==2019.4.13
venusian==3.0.0
WebOb==1.8.6
websocket-client==0.57.0
wrapt==1.12.1
zope.deprecation==4.4.0
zope.interface==5.1.0
